Proyecto terminado

+++ js/controlador/controlador.js
@@ -6,18 +6,73 @@ var Controlador = function (modelo) {
 };
 
 Controlador.prototype = {
+
   agregarPregunta: function (pregunta, respuestas) {
-    this.modelo.agregarPregunta(pregunta, respuestas);
+    if (pregunta !== '' && respuestas !== '') {
+      this.modelo.agregarPregunta(pregunta, respuestas);
+    } else {
+      Swal.fire({
+        icon: 'error',
+        title: 'Error...',
+        text: 'You cannot register an empty question!',
+      });
+    }
   },
+
   borrarPregunta: function (idPregunta) {
-    this.modelo.borrarPregunta(idPregunta);
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.modelo.borrarPregunta(idPregunta);
+        Swal.fire(
+          'Deleted!',
+          'Your question has been deleted.',
+          'success'
+        )
+      }
+    })
   },
+
   deleteAllQuestions: function () {
-    this.modelo.deleteAllQuestions();
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        this.modelo.deleteAllQuestions();
+        Swal.fire(
+          'Deleted!',
+          'Your questions has been deleted.',
+          'success'
+        )
+      }
+    })
   },
+
   agregarVoto: function (pregunta, respuestaSeleccionada) {
-    this.modelo.agregarVoto(pregunta, respuestaSeleccionada);
+    if (pregunta !== undefined && respuestaSeleccionada !== undefined) {
+      this.modelo.agregarVoto(pregunta, respuestaSeleccionada);
+    } else {
+      Swal.fire({
+        icon: 'error',
+        title: 'Error...',
+        text: 'Select an option before sending the reply !!',
+      });
+    }
   },
+
   editQuestions: function (idPregunta) {
     this.modelo.editQuestions(idPregunta);
   }
